+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ### Added
 
+- Add a `--minimal-update` flag to `lock` to generate a lockfile
+  with minimum dependency changes from a previous lockfile. (#<PR_NUMBER>,
+  @NathanReb)
+
 ### Changed
 
 ### Deprecated